labs/lab02-code-analyzer-agent/python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.5
anthropic==0.18.0
openai==1.12.0
google-generativeai==0.3.0
python-dotenv==1.0.0
